requirements.txt
crewai>=0.28.8
langchain-groq>=0.1.0
langchain-core>=0.1.0
feedparser>=6.0.10
python-dotenv>=1.0.0
pydantic>=2.0.0
aiohttp>=3.8.0
fastapi
uvicorn
requests
beautifulsoup4